search_faiss/requirements.txt
faiss-cpu
numpy
openai
dotenv